Move render unit tests out of the integration tests module
